drivers: display: Add GD32 TLI (TFT-LCD) controller support
Add GD32 TLI (TFT-LCD Interface) display controller driver and device tree
bindings for GD32F527. This includes Kconfig, CMake, driver source, and
devicetree overlays for the GD32F527I-EVAL board.

+++ drivers/display/display_gd32_tli.c
@@ -65,7 +65,7 @@ struct display_gd32_tli_config {
 	uint16_t height;
 	uint16_t pllsai_n;
 	uint16_t pllsai_r;
-	uint8_t pllsair_div;
+	uint32_t pllsair_div;
 };
 
 struct display_gd32_tli_data {
